added policy number to dance step policy so that we can branch on it in compiled tensorflow XLA code (which cannot branch on string constants) fixed physics of cart pole potential and kinetic energy to put in mass  and length and correct constants renamed cartpole cost function to cartpole_dancer_cost.py enhanced the load_or_reload_config_if_modified function to set a policy_number variable if the property name ends in an int
