+++ Control_Toolkit_ASF/Cost_Functions/CartPole/cartpole_dance.csv
@@ -1,16 +1,17 @@
 # cartpole dancer dance sequence
+# policies can be balance1, spin2, shimmy3, cartonly4
 duration, cartpos, policy, option, freq, amp, freq2, amp2
-3,  0, balance, up,,,,
-8, .1, spin, cw, 5,,,
-#1,  0, balance, up,,,,
-8, -.1,  spin,ccw, 5,,,
-2, -.1, balance, up,,,,
-10, 0, shimmy,up, 2, .05, 1, .1
-1,  0, balance, up,,,,
-10, 0, shimmy,up, 3, .001, 3, .1
-1.5, .15, balance, up,,,,
-1.5, -.15, balance, up,,,,
-#2, 0, balance, up,,,,
-# 4, 0, cartonly,.9, 2, .1,,
-3, 0, balance, down,,,,
+3,  0, balance1, up,,,,
+8, .1, spin2, cw, 5,,,
+#1,  0, balance1, up,,,,
+8, -.1,  spin2,ccw, 5,,,
+2, -.1, balance1, up,,,,
+10, 0, shimmy3,up, 2, .05, 1, .1
+1,  0, balance1, up,,,,
+10, 0, shimmy3,up, 3, .001, 3, .1
+1.5, .15, balance1, up,,,,
+1.5, -.15, balance1, up,,,,
+#2, 0, balance1, up,,,,
+# 4, 0, cartonly4,.9, 2, .1,,
+3, 0, balance1, down,,,,
